# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: FactorSage API
  version: "1.0.0"
  summary: Stock research, strategy design, backtesting and monitoring.
  description: |
    The HTTP API behind FactorSage.

    **Authentication** is a session cookie issued by `POST /auth/login` or by the Google OAuth
    callback. It is `HttpOnly`, so browser clients send it automatically and never read it; a
    non-browser client must store and resend it. There is no bearer token and no API key.

    **Authorization** has two independent axes, both resolved server-side from persisted state and
    never from anything the client sends: a role (`USER` / `ADMIN`) and a commercial plan
    (`FREE` / `STARTER` / `PRO`). A plan limit is refused with `403` and a stable
    `ENTITLEMENT_*` code. Ownership is not a separate axis — every user-owned collection is scoped
    to the caller, and somebody else's resource answers exactly like one that does not exist.

    **Rate limiting** is applied per endpoint class from a central policy catalog, keyed by user id
    for authenticated traffic and by client IP otherwise. Exceeding a policy answers `429` with
    `Retry-After`. See `info.x-rate-limit-policies`.

    **Errors** share one machine-readable shape: an HTTP `statusCode`, a human `message`, and — for
    every refusal a client may reasonably branch on — a stable `code`. Clients should branch on
    `code`, never on `message`.
  contact:
    name: FactorSage engineering
  license:
    name: UNLICENSED

...

...
servers:
...
paths:
...
  /monitors:
    get:
      tags: [Monitors]
      operationId: listMonitors
      summary: The caller's monitors.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: Monitor summaries.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MonitorSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    post:
      tags: [Monitors]
      operationId: createMonitor
      summary: Create a monitor.
      description: |
        A monitor is a live strategy reference plus a live stock list reference plus `enabled`, and
        nothing else. There is no pinned version, no execution parameter and no cadence: scan
        frequency is an application decision. Unknown fields are **rejected** rather than ignored,
        so an attempt to set a cadence fails loudly instead of appearing to work.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: monitor-mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMonitorRequest" }
      responses:
        "201":
          description: The created monitor, with the current status of every monitored security.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonitorDetail" }
        "400":
          description: Malformed body, an unknown field, or a strategy or list the caller does not own.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/EntitlementDenied" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /monitors/{monitorId}:
    parameters:
      - { $ref: "#/components/parameters/MonitorId" }
    get:
      tags: [Monitors]
      operationId: getMonitor
      summary: One monitor, its monitored securities and its newest signals.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: The monitor.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonitorDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    patch:
      tags: [Monitors]
      operationId: updateMonitor
      summary: Rename, enable/disable, or rebind a monitor.
      description: |
        Changing `strategyId` or `stockListId` **rebinds** the monitor. That is a configuration
        boundary, not an edit: transition state is discarded, active signals are resolved and
        `lastScanAt` is cleared, while signal history is kept. An in-flight scan cycle is fenced off
        so it cannot write state belonging to the old configuration.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: monitor-mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMonitorRequest" }
      responses:
        "200":
          description: The updated summary.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonitorSummary" }
        "400":
          description: Malformed body, an unknown field, or a strategy or list the caller does not own.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/EntitlementDenied" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

    delete:
      tags: [Monitors]
      operationId: deleteMonitor
      summary: Delete a monitor.
      description: Signals it produced are append-only durable records and are never deleted with it.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: monitor-mutation }
      responses:
        "204":
          description: Deleted.
          headers: *rateLimitHeaders
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/OwnedResourceNotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ---------------------------------------------------------------------------------------------
  # Billing
  # ---------------------------------------------------------------------------------------------
  /billing/status:
    get:
      tags: [Billing]
      operationId: getBillingStatus
      summary: The caller's subscription state and the price catalog.
      description: |
        `plan` is the **persisted** plan — the same column every entitlement check reads — and is
        deliberately not recomputed from the payment provider for this response. Showing something
        the guards would refuse is exactly the optimistic display the billing decision forbids.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: standard-read }
      responses:
        "200":
          description: Billing status. `subscription` is null when there is none.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /billing/refresh:
    post:
      tags: [Billing]
      operationId: refreshBillingStatus
      summary: Reconcile from the payment provider, then return status.
      description: |
        What the billing page calls after returning from a hosted provider surface. It is a read of
        authoritative provider state run through the same reconciliation a webhook runs — not a
        grant — so a cancelled subscription or a failed card produces the honest answer rather than
        the optimistic one. Its value is latency only; webhooks remain the authority, and a
        reconciliation failure still returns the persisted status rather than failing the page.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: billing-mutation }
      responses:
        "200":
          description: Billing status after reconciliation.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ThrottlingUnavailable" }
  /billing/checkout:
    post:
      tags: [Billing]
      operationId: startCheckout
      summary: Start hosted checkout for one catalog price.
      description: |
        Accepts one of four **logical** price keys and nothing else. No provider price, customer or
        subscription identifier is accepted from a client, anywhere in this API.

        Returns the URL rather than redirecting, so the browser controls navigation and a refusal
        stays a normal JSON error.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: billing-mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingTargetRequest" }
      responses:
        "201":
          description: A hosted checkout session URL.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingCheckoutResponse" }
        "400":
          description: "Not a known price key. `code: BILLING_INVALID_PRICE_KEY`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: "Already subscribed. `code: BILLING_ALREADY_SUBSCRIBED`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/BillingUnavailable" }
  /billing/portal:
    post:
      tags: [Billing]
      operationId: openBillingPortal
      summary: Open the hosted customer portal.
      description: |
        Bound to the session's user; there is no parameter naming a customer, which is what makes
        "one user cannot open another's portal" a property of the shape rather than of a check.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: billing-mutation }
      responses:
        "201":
          description: A hosted portal session URL.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingPortalResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: "No subscription to manage. `code: BILLING_NO_SUBSCRIPTION`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/BillingUnavailable" }
  /billing/change:
    post:
      tags: [Billing]
      operationId: changeSubscription
      summary: Move an existing subscription to another catalog price.
      description: |
        Exists because the hosted portal cannot express immediate-for-upgrades and
        period-end-for-downgrades at the same time. The response says which effect was applied and
        when it takes hold. An upgrade whose payment has not succeeded never grants a plan.
      security: [{ cookieAuth: [] }]
      x-rate-limit: { policy: billing-mutation }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BillingTargetRequest" }
      responses:
        "201":
          description: The transition that was applied or scheduled.
          headers: *rateLimitHeaders
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingChangeResponse" }
        "400":
          description: "Not a known price key. `code: BILLING_INVALID_PRICE_KEY`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402":
          description: "Payment is required before the change can apply. `code: BILLING_PAYMENT_REQUIRED`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingError" }
        "409":
          description: |
            The transition is not allowed, there is nothing to change, or no subscription exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BillingError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/BillingUnavailable" }

  # ---------------------------------------------------------------------------------------------
  # Admin
  # ---------------------------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: intrinsic_auth
      description: |
        The session cookie issued by `POST /auth/login` or by the Google callback. `HttpOnly`,
        `SameSite=Lax`, and `Secure` in production. Its name is deployment configuration
        (`AUTH_COOKIE_NAME`); `intrinsic_auth` is the default. A browser sends it automatically —
        cross-origin callers must send credentials explicitly, and the deployment's CORS allowlist
        must contain their origin.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: |
        Signature over the raw request body, verified with the webhook signing secret. The only
        authentication the webhook endpoint has, and the only one it needs.
  headers:
    RateLimitPolicy:
      description: The rate-limit policy that produced the numbers in the other headers.
      schema: { type: string }
      example: standard-read
...
  parameters:
...
    MonitorId:
      name: monitorId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: No session, or a session that is expired, forged or otherwise unusable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    EntitlementDenied:
      description: |
        The caller's plan does not allow this. The body carries a stable `ENTITLEMENT_*` code plus
        the numbers behind it, so a client can say "your plan allows 10, this change would make 11"
        without parsing the message.

        This is a **product limit**, not a rate limit: retrying identically will fail identically,
        and the remedy is a different plan or a smaller request — never waiting.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/EntitlementError" }
    OwnedResourceNotFound:
      description: |
        No such resource for this caller. A resource that exists but belongs to somebody else
        answers exactly like one that does not exist, so this response reveals nothing about another
        user's data.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    BillingUnavailable:
      description: |
        The payment provider or this deployment's billing configuration is the problem, not the
        caller. Deliberately not a `500`: the request was fine and a retry after an operator fixes
        configuration will work.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BillingError" }
    TooManyRequests:
      description: |
        The caller exceeded this endpoint's rate-limit policy. `Retry-After` is the measured time
        until the allowance refills; waiting exactly that long is sufficient, and retrying sooner is
        refused again without consuming anything further.

        A client should surface this as throttling with the wait, never as a generic failure — and
        on `POST /auth/login` in particular it must not be shown as "wrong credentials", because the
        attempt is refused whether or not the password was correct.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        RateLimit-Policy: { $ref: "#/components/headers/RateLimitPolicy" }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitError" }
          examples:
            search:
              value:
                statusCode: 429
                message: "Too many requests. Retry in 42 seconds."
                code: RATE_LIMITED
                policy: stock-search
                retryAfterSeconds: 42
    ThrottlingUnavailable:
      description: |
        The rate-limit store could not be reached and this endpoint's policy is fail-closed, so the
        request was refused rather than served unmetered. Documented only on the security- and
        payment-sensitive endpoints, because every other policy fails open.

        The caller's allowance is untouched. A client must not present this as "you are going too
        fast"; it carries its own code for exactly that reason.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RateLimitError" }
          examples:
            unavailable:
              value:
                statusCode: 503
                message: "Request throttling is temporarily unavailable. Please retry shortly."
                code: RATE_LIMIT_UNAVAILABLE
                policy: auth-sensitive
                retryAfterSeconds: 5
  schemas:
...
    Error:
      type: object
      description: |
        The API's error shape. `code` is present on every refusal a client can reasonably branch on
        and absent on generic framework errors; branch on it, never on `message`.
      required: [statusCode, message]
      properties:
        statusCode: { type: integer, description: Repeats the HTTP status. }
        message:
          {
            type: string,
            description: Human-readable. Never parsed by a client.,
          }
        code:
          {
            type: string,
            description: Stable machine-readable reason,
            when there is one.,
          }
      examples:
        - { statusCode: 404, message: "Stock symbol was not found" }
...
    BillingError:
      allOf:
        - { $ref: "#/components/schemas/Error" }
        - type: object
          required: [code]
          properties:
            code:
              type: string
              description: |
                Stable billing reason. Carries nothing about the payment provider — no provider
                error object, request id, customer or subscription identifier — because the original
                failure is logged server-side before translation.
              enum:
                - BILLING_AUTH_REQUIRED
                - BILLING_INVALID_PRICE_KEY
                - BILLING_PAYMENT_REQUIRED
                - BILLING_ALREADY_SUBSCRIBED
                - BILLING_NO_SUBSCRIPTION
                - BILLING_CHANGE_NOT_ALLOWED
                - BILLING_CUSTOMER_CONFLICT
                - BILLING_SUBSCRIPTION_CONFLICT
                - BILLING_WEBHOOK_INVALID_SIGNATURE
                - BILLING_WEBHOOK_UNSUPPORTED_STATE
                - BILLING_NOT_CONFIGURED
                - BILLING_CATALOG_MISCONFIGURED
                - BILLING_STRIPE_UNAVAILABLE
                - BILLING_PORTAL_UNAVAILABLE
                - BILLING_CHECKOUT_FAILED
                - BILLING_SUBSCRIPTION_UPDATE_FAILED
            priceKey: { $ref: "#/components/schemas/BillingPriceKey" }
            plan: { $ref: "#/components/schemas/UserPlan" }
...
    NullableDateTime:
      oneOf: [{ type: string, format: date-time }, { type: "null" }]
...
    MonitorOperationalStatus:
      type: string
      enum: [ACTIVE, PAUSED, BLOCKED]
      description: |
        Derived, never persisted as user intent. `enabled` is what the user set; a monitor may still
        be `BLOCKED` because the caller's plan no longer covers it, and re-subscribing unblocks it
        without the user having to re-enable anything.
    MonitorBlockedReason:
      type: string
      enum: [PLAN_MONITOR_LIMIT, PLAN_LIST_SYMBOL_LIMIT]
...
    MonitorSummary:
      type: object
      required:
        [
          id,
          name,
          enabled,
          strategyId,
          strategyName,
          stockListId,
          stockListName,
          securityCount,
          activeSignalCount,
          createdAt,
          updatedAt,
          operationalStatus,
        ]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, maxLength: 120 }
        enabled:
          {
            type: boolean,
            description: The user's intent. Never rewritten by the server.,
          }
        strategyId: { type: string, format: uuid }
        strategyName: { type: string }
        stockListId: { type: string, format: uuid }
        stockListName: { type: string }
        securityCount: { type: integer, minimum: 0 }
        activeSignalCount: { type: integer, minimum: 0 }
        lastScanAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        operationalStatus:
          { $ref: "#/components/schemas/MonitorOperationalStatus" }
        blockedReason: { $ref: "#/components/schemas/MonitorBlockedReason" }
    MonitorDetail:
      allOf:
        - { $ref: "#/components/schemas/MonitorSummary" }
        - type: object
          required: [securities, signals]
          properties:
            securities:
              type: array
              description: Every security in the current list with its status, ordered by symbol.
              items: { $ref: "#/components/schemas/MonitorSecurityEvaluation" }
            signals:
              type: array
              description: The newest signals, newest first.
              items: { $ref: "#/components/schemas/MonitorSignal" }
    CreateMonitorRequest:
      type: object
      additionalProperties: false
      description: |
        Unknown properties are rejected rather than ignored — that is what stops an `interval`,
        `cadence` or `schedule` field from ever looking like it was accepted.
      required: [name, strategyId, stockListId]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        strategyId: { type: string, format: uuid }
        stockListId: { type: string, format: uuid }
        enabled: { type: boolean, default: true }
    UpdateMonitorRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      description: At least one property; unknown properties are rejected.
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        enabled: { type: boolean }
        strategyId:
          { type: string, format: uuid, description: Rebinds the monitor. }
        stockListId:
          { type: string, format: uuid, description: Rebinds the monitor. }

    # -------------------------------------------------------------------------------------------
    # Billing
    # -------------------------------------------------------------------------------------------
...
    BillingStatus:
      type: object
      required:
        [
          plan,
          billingEnabled,
          subscription,
          canStartCheckout,
          canOpenPortal,
          canChangePlan,
          catalog,
        ]
      properties:
        plan: { $ref: "#/components/schemas/UserPlan" }
        billingEnabled:
          {
            type: boolean,
            description: Whether this deployment has billing configured at all.,
          }
        subscription:
          oneOf:
            - { $ref: "#/components/schemas/BillingSubscriptionView" }
            - { type: "null" }
        canStartCheckout: { type: boolean }
        canOpenPortal: { type: boolean }
        canChangePlan: { type: boolean }
        catalog:
          type: array
          items: { $ref: "#/components/schemas/BillingCatalogEntry" }
    BillingTargetRequest:
      type: object
      required: [priceKey]
      properties:
        priceKey: { $ref: "#/components/schemas/BillingPriceKey" }
    BillingCheckoutResponse:
      type: object
      required: [checkoutUrl]
      properties:
        checkoutUrl: { type: string, format: uri }
    BillingPortalResponse:
      type: object
      required: [portalUrl]
      properties:
        portalUrl: { type: string, format: uri }
    BillingChangeResponse:
      type: object
      required: [effect, kind, effectiveAt, plan]
      properties:
        effect:
          type: string
          enum: [IMMEDIATE, PERIOD_END]
          description: Upgrades apply immediately; downgrades apply at the period end.
        kind: { type: string, enum: [UPGRADE, DOWNGRADE, INTERVAL_CHANGE] }
        effectiveAt: { $ref: "#/components/schemas/NullableDateTime" }
        plan: { $ref: "#/components/schemas/UserPlan" }

    # -------------------------------------------------------------------------------------------
    # Operations
    # -------------------------------------------------------------------------------------------
...